+++ src/FMBot.Bot/Services/WhoKnows/CrownService.cs
@@ -2,14 +2,19 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Dapper;
 using Discord.WebSocket;
+using FMBot.Bot.Configurations;
 using FMBot.Bot.Extensions;
 using FMBot.Bot.Models;
 using FMBot.Domain;
+using FMBot.Domain.Models;
 using FMBot.LastFM.Services;
 using FMBot.Persistence.Domain.Models;
 using FMBot.Persistence.EntityFrameWork;
 using Microsoft.EntityFrameworkCore;
+using Npgsql;
+using PostgreSQLCopyHelper;
 
 namespace FMBot.Bot.Services.WhoKnows
 {
@@ -17,11 +22,13 @@ namespace FMBot.Bot.Services.WhoKnows
     {
         private readonly IDbContextFactory<FMBotDbContext> _contextFactory;
         private readonly LastFmService _lastFmService;
+        private readonly GlobalUpdateService _globalUpdateService;
 
-        public CrownService(IDbContextFactory<FMBotDbContext> contextFactory, LastFmService lastFmService)
+        public CrownService(IDbContextFactory<FMBotDbContext> contextFactory, LastFmService lastFmService, GlobalUpdateService globalUpdateService)
         {
             this._contextFactory = contextFactory;
             this._lastFmService = lastFmService;
+            this._globalUpdateService = globalUpdateService;
         }
 
         public async Task<CrownModel> GetAndUpdateCrownForArtist(IList<WhoKnowsObjectWithUser> users, Persistence.Domain.Models.Guild guild, string artistName)
@@ -71,6 +78,7 @@ namespace FMBot.Bot.Services.WhoKnows
                 {
                     currentCrownHolder.CurrentPlaycount = topUser.Playcount;
                     currentCrownHolder.Modified = DateTime.UtcNow;
+                    currentCrownHolder.SeededCrown = false;
 
                     db.Entry(currentCrownHolder).State = EntityState.Modified;
                     await db.SaveChangesAsync();
@@ -91,7 +99,7 @@ namespace FMBot.Bot.Services.WhoKnows
             if (currentCrownHolder != null && topUser.UserId != currentCrownHolder.UserId)
             {
                 var currentPlaycountForCrownHolder =
-                    await GetCurrentPlaycountForUser(artistName, currentCrownHolder.User.UserNameLastFM);
+                    await GetCurrentPlaycountForUser(artistName, currentCrownHolder.User.UserNameLastFM, currentCrownHolder.UserId);
 
                 if (currentPlaycountForCrownHolder == null)
                 {
@@ -116,11 +124,6 @@ namespace FMBot.Bot.Services.WhoKnows
                     };
                 }
 
-                if (currentCrownHolder.CurrentPlaycount != currentPlaycountForCrownHolder)
-                {
-                    currentCrownHolder.CurrentPlaycount = topUser.Playcount;
-                }
-
                 currentCrownHolder.Active = false;
                 currentCrownHolder.Modified = DateTime.UtcNow;
 
@@ -135,17 +138,21 @@ namespace FMBot.Bot.Services.WhoKnows
                     Created = DateTime.UtcNow,
                     Modified = DateTime.UtcNow,
                     StartPlaycount = topUser.Playcount,
-                    CurrentPlaycount = topUser.Playcount
+                    CurrentPlaycount = topUser.Playcount,
+                    SeededCrown = false
                 };
 
                 await db.UserCrowns.AddAsync(newCrown);
 
                 await db.SaveChangesAsync();
 
+                var currentCrownHolderName = users.FirstOrDefault(f => f.UserId == currentCrownHolder.UserId)?.DiscordName;
+
                 return new CrownModel
                 {
                     Crown = newCrown,
-                    CrownResult = $"Crown stolen by {topUser.DiscordName}! (From **{currentCrownHolder.CurrentPlaycount}** to **{topUser.Playcount}** plays)."
+                    CrownResult = $"Crown stolen by {topUser.DiscordName} with `{topUser.Playcount}` plays! \n" +
+                                  $"*Previous owner: {currentCrownHolderName ?? currentCrownHolder.User.UserNameLastFM} with `{currentCrownHolder.CurrentPlaycount}` plays*."
                 };
             }
 
@@ -165,7 +172,8 @@ namespace FMBot.Bot.Services.WhoKnows
                         Created = DateTime.UtcNow,
                         Modified = DateTime.UtcNow,
                         StartPlaycount = topUser.Playcount,
-                        CurrentPlaycount = topUser.Playcount
+                        CurrentPlaycount = topUser.Playcount,
+                        SeededCrown = false
                     };
 
                     await db.UserCrowns.AddAsync(newCrown);
@@ -193,14 +201,93 @@ namespace FMBot.Bot.Services.WhoKnows
             return null;
         }
 
-        private async Task<long?> GetCurrentPlaycountForUser(string artistName, string lastFmUserName)
+        public async Task<int> SeedCrownsForGuild(Persistence.Domain.Models.Guild guild, IList<UserCrown> existingCrowns)
+        {
+            const string sql = "SELECT DISTINCT ON(ua.name) " +
+                                   "ua.user_id, " +
+                                   "ua.name, " +
+                                   "ua.playcount " +
+                                   "FROM user_artists AS ua " +
+                                   "INNER JOIN users AS u ON ua.user_id = u.user_id " +
+                                   "INNER JOIN guild_users AS gu ON gu.user_id = u.user_id " +
+                                   "WHERE gu.guild_id = @guildId AND playcount > @minPlaycount " +
+                                   "AND NOT ua.user_id = ANY(SELECT user_id FROM guild_blocked_users WHERE blocked_from_crowns = true AND guild_id = @guildId) " +
+                                   "ORDER BY ua.name, ua.playcount DESC;";
+
+            DefaultTypeMap.MatchNamesWithUnderscores = true;
+            await using var connection = new NpgsqlConnection(ConfigData.Data.Database.ConnectionString);
+            await connection.OpenAsync();
+
+            var minPlaycount = guild.CrownsMinimumPlaycountThreshold ?? Constants.DefaultPlaysForCrown;
+
+            var topUsersForPlaycount = await connection.QueryAsync<CrownSeedDto>(sql, new
+            {
+                guild.GuildId,
+                minPlaycount
+            });
+
+            var now = DateTime.UtcNow;
+            var crownsToSeed = topUsersForPlaycount.Select(s =>
+                new UserCrown
+                {
+                    UserId = s.UserId,
+                    Active = true,
+                    ArtistName = s.Name,
+                    Created = existingCrowns.FirstOrDefault(f => f.UserId == s.UserId && f.ArtistName.ToLower() == s.Name.ToLower())?.Created ?? now,
+                    Modified = now,
+                    CurrentPlaycount = s.Playcount,
+                    StartPlaycount = existingCrowns.FirstOrDefault(f => f.UserId == s.UserId && f.ArtistName.ToLower() == s.Name.ToLower())?.StartPlaycount ?? s.Playcount,
+                    GuildId = guild.GuildId,
+                    SeededCrown = true
+                }).ToList();
+
+            if (existingCrowns != null && existingCrowns.Any())
+            {
+                crownsToSeed = crownsToSeed.Where(w =>
+                        !existingCrowns
+                            .Where(wh => !wh.SeededCrown)
+                            .Select(s => s.ArtistName.ToLower())
+                            .Contains(w.ArtistName.ToLower()))
+                    .ToList();
+            }
+            const string deleteOldSeededCrownsSql = "DELETE FROM public.user_crowns " +
+                                                    "WHERE guild_id = @guildId AND seeded_crown = true;";
+
+            await connection.ExecuteAsync(deleteOldSeededCrownsSql, new
+            {
+                guild.GuildId,
+            });
+
+            var copyHelper = new PostgreSQLCopyHelper<UserCrown>("public", "user_crowns")
+                .MapInteger("guild_id", x => x.GuildId)
+                .MapInteger("user_id", x => x.UserId)
+                .MapText("artist_name", x => x.ArtistName)
+                .MapInteger("current_playcount", x => x.CurrentPlaycount)
+                .MapInteger("start_playcount", x => x.StartPlaycount)
+                .MapTimeStamp("created", x => x.Created)
+                .MapTimeStamp("modified", x => x.Created)
+                .MapBoolean("active", x => x.Active)
+                .MapBoolean("seeded_crown", x => x.SeededCrown);
+
+            await copyHelper.SaveAllAsync(connection, crownsToSeed);
+            await connection.CloseAsync();
+
+            return crownsToSeed.Count;
+        }
+
+        private async Task<long?> GetCurrentPlaycountForUser(string artistName, string lastFmUserName, int userId)
         {
             var artist = await this._lastFmService.GetArtistInfoAsync(artistName, lastFmUserName);
 
-            if (!artist.Success)
+            await this._globalUpdateService.UpdateUser(new UpdateUserQueueItem(userId, 0));
+
+            if (!artist.Success || !artist.Content.Artist.Stats.Userplaycount.HasValue)
             {
                 return null;
             }
+
+            await this._globalUpdateService.CorrectUserArtistPlaycount(userId, artistName,
+                artist.Content.Artist.Stats.Userplaycount.Value);
 
             return artist.Content.Artist.Stats.Userplaycount;
         }
@@ -213,7 +300,7 @@ namespace FMBot.Bot.Services.WhoKnows
                 .Include(i => i.User)
                 .OrderByDescending(o => o.CurrentPlaycount)
                 .Where(f => f.GuildId == guild.GuildId &&
-                            EF.Functions.ILike(f.ArtistName, artistName))
+                            f.ArtistName.ToLower() == artistName.ToLower())
                 .ToListAsync();
         }
 
@@ -266,12 +353,35 @@ namespace FMBot.Bot.Services.WhoKnows
                 .CountAsync();
         }
 
+        public async Task<IList<UserCrown>> GetAllCrownsForGuild(int guildId)
+        {
+            await using var db = this._contextFactory.CreateDbContext();
+            return await db.UserCrowns
+                .AsQueryable()
+                .Include(i => i.User)
+                .Where(w => w.GuildId == guildId)
+                .ToListAsync();
+        }
+
         public async Task RemoveAllCrownsFromGuild(Persistence.Domain.Models.Guild guild)
         {
             await using var db = this._contextFactory.CreateDbContext();
             var guildCrowns = await db.UserCrowns
                 .AsQueryable()
-                .Where(f => f.GuildId == guild.GuildId)
+                .Where(w => w.GuildId == guild.GuildId)
+                .ToListAsync();
+
+            db.UserCrowns.RemoveRange(guildCrowns);
+
+            await db.SaveChangesAsync();
+        }
+
+        public async Task RemoveAllSeededCrownsFromGuild(Persistence.Domain.Models.Guild guild)
+        {
+            await using var db = this._contextFactory.CreateDbContext();
+            var guildCrowns = await db.UserCrowns
+                .AsQueryable()
+                .Where(w => w.GuildId == guild.GuildId && w.SeededCrown)
                 .ToListAsync();
 
             db.UserCrowns.RemoveRange(guildCrowns);
@@ -317,9 +427,12 @@ namespace FMBot.Bot.Services.WhoKnows
                                 f.GuildId == guild.GuildId)
                     .ToListAsync();
 
-                db.UserCrowns.RemoveRange(userGuildCrowns);
+                if (userGuildCrowns != null && userGuildCrowns.Any())
+                {
+                    db.UserCrowns.RemoveRange(userGuildCrowns);
 
-                await db.SaveChangesAsync();
+                    await db.SaveChangesAsync();
+                }
             }
         }
